_scripts/requirements.txt
astroid==2.11.5
autopep8==1.6.0
dill==0.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pep8==1.7.1
platformdirs==2.5.2
plum-py==0.8.0
pycodestyle==2.8.0
pylint==2.13.9
PyYAML==6.0
toml==0.10.2
tomli==2.0.1
wrapt==1.14.1
apsw==3.40.1.0
